# Introduce performance benchmark