--- a/astronoby.gemspec
+++ b/astronoby.gemspec
@@ -32,6 +32,8 @@ Gem::Specification.new do |spec|
   spec.add_dependency "ephem", "~> 0.3"
   spec.add_dependency "matrix", "~> 0.4.2"
 
+  spec.add_development_dependency "benchmark", "~> 0.4"
+  spec.add_development_dependency "irb", "~> 1.14"
   spec.add_development_dependency "rake", "~> 13.0"
   spec.add_development_dependency "rspec", "~> 3.0"
   spec.add_development_dependency "rubyzip", "~> 2.3"